Add AJAX calls for booking slots view
I've added AJAX functionality to the view bookings slot page but I'm getting an empty array in return so the timeslot buttons will appear empty. Not sure if this is an issue with backend not being able to retrieve the date from datepicker or that there are no available time slots in the database.